fix(cv610): audio failure must not take video down, and report it honestly
Pre-merge review of the widened capability surface. Making audio.enabled an
HTTP control exposed a path that could kill the daemon: cv610_init() treated
a failed cv610_audio_start() as fatal, so setting audio.enabled=true on a
video-only boot (CV610_AUDIO=0, acodec modules absent) would respawn into a
dead daemon and take video with it. docs/CV610_BACKEND.md already warned
this was why the audio fields were not advertised.

Star6E is the precedent: star6e_pipeline.c discards the audio init result.
Do the same here -- warn, name the cause, and run without audio.

Also: /api/v1/audio/status reported "running":true whenever the state
existed. The capture thread exits by break on select or SDK errors without
clearing its flag, so a dead thread read as running forever. The thread now
clears the flag on every exit path, which makes it mean "capture thread
alive", and cv610_audio_is_running() exposes it.

Verified on the demo board by reproducing the exact regression: CV610_AUDIO=0
with audio.enabled=true left in config. Daemon starts, logs "audio did not
start; continuing without it", video streams at 100.1 fps, and
/api/v1/audio/status reports enabled:false. Restored to CV610_AUDIO=1: audio
back, counters advancing, 0 drops.

Also confirmed the widened allowlist cannot reach an unsafe path --
validate_backend_config() runs cv610_validate_config() on the staged config
inside stage_params_into_cfg(), before persist or respawn. Probed on device:

  video0.fps=45          -> 409 "must be 30, 60, 90, or 100"
  video0.size=1280x720   -> 409 "requires video0 1920x1080"
  outgoing.server=shm:// -> 409 "supports udp://, unix://, or frame-shm://"
  outgoing.enabled=false -> 409 "audio requires outgoing output enabled"
  video0.rc_mode=vbr     -> 501 not supported on this backend

with the on-disk config unchanged after all five. shm:// matters most: it
parses cleanly and would have failed later in cv610_output_start().

Comment cleanup per CONVENTIONS.md section 9 (why, not what), and the
in-tree docs and boot scripts updated for the runtime sensor clock:
CV610_SENSOR_PROFILE is now a fallback, not the thing that picks the mode.

Host test-werror 2479/0.

+++ src/cv610_pipeline.c
@@ -99,36 +99,26 @@ static void *isp_thread_fn(void *arg)
 
 /* ------------------------------------------------------------------ MIPI -- */
 
-/* Sensor MCLK, chosen per mode.
- *
- * The IMX662 driver selects INCK_SEL per mode — 0x04 for the 100 fps path,
- * which feeds 27 MHz while the sensor believes it is on its 24 MHz profile,
- * and 0x01 for 30/60/90 on 37.125 MHz (see
- * sensors/cv610/imx662/imx662_sensor_ctl.c, imx662_default_reg_init()).  The
- * SoC-side clock has to match that choice or the sensor keeps its programmed
- * line timing against the wrong input clock and the frame rate scales by the
- * clock ratio — measured 43.6 fps for a 60 fps mode on the 27 MHz clock
- * (60 x 27/37.125), with every status surface still reporting 60.
- *
- * The clock lives in a CRG register that only the kernel can write, and the
- * vendor MIPI ioctls only gate it (ENABLE/DISABLE_SENSOR_CLOCK) rather than
- * set a rate, so sys_config exposes it as sns0_clk_hz.  We ask for a
- * frequency and let the module own the register encoding. */
+/* The IMX662 needs 37.125 MHz for 30/60/90 fps and 27 MHz for 100 fps, which
+ * takes the sensor's 24 MHz INCK profile overclocked (INCK_SEL 0x04, see
+ * imx662_default_reg_init() in sensors/cv610/imx662).  A mismatch is silent:
+ * the sensor keeps its programmed line timing against the wrong input clock
+ * and the rate scales by the ratio — measured 43.6 fps for a 60 fps mode on
+ * the 27 MHz clock.  The clock is a CRG register only the kernel can write,
+ * and the MIPI ioctls gate it without setting a rate, so sys_config exposes
+ * it as a frequency and keeps the register encoding. */
 #define CV610_SNS0_CLK_HZ_PATH "/sys/module/open_sys_config/parameters/sns0_clk_hz"
 
 static void sensor_clock_select(uint32_t fps)
 {
-	/* Mirrors imx662_default_reg_init(): only the 100 fps mode takes the
-	 * overclock path.  Keep the two in step. */
 	const unsigned hz = (fps > 90) ? 27000000u : 37125000u;
 	char buf[16];
 	int len, fd;
 
 	fd = open(CV610_SNS0_CLK_HZ_PATH, O_WRONLY);
 	if (fd < 0) {
-		/* An older sys_config has no runtime knob; the clock is then
-		 * whatever CV610_SENSOR_PROFILE selected at insmod, which is
-		 * right for exactly one mode.  Say which one this needs. */
+		/* Older sys_config: the clock stays as CV610_SENSOR_PROFILE left
+		 * it, which suits exactly one mode. */
 		fprintf(stderr,
 			"WARNING: %s absent — sensor clock stays as loaded; "
 			"%u fps needs %u Hz and will otherwise run at the wrong rate\n",
@@ -136,7 +126,7 @@ static void sensor_clock_select(uint32_t fps)
 		return;
 	}
 	len = snprintf(buf, sizeof(buf), "%u", hz);
-	if (len < 0 || (size_t)len >= sizeof(buf) ||
+	if (len < 0 || (size_t)len >= (int)sizeof(buf) ||
 		write(fd, buf, (size_t)len) != len) {
 		fprintf(stderr, "WARNING: set sensor clock %u Hz: %s\n", hz,
 			strerror(errno));
